+++ Exp2_S5/src/main/java/com/duoc/exp2_s5/ui/admin/MenuAsistente.java
@@ -22,7 +22,6 @@ public class MenuAsistente {
             System.out.println("\n===== MENÚ ASISTENTE =====");
             System.out.println("1. Agregar libro");
             System.out.println("2. Editar stock de libro");
-            System.out.println("3. Bloquear/Desbloquear usuario");
             System.out.println("0. Volver al menú principal");
             System.out.print("Seleccione una opción: ");
             
@@ -39,14 +38,11 @@ public class MenuAsistente {
                 case 2:
                     new ModificarCantidadLibroMenu(svc, sc).mostrar();
                     break;
-                case 3:
-                    System.out.println("→ Aquí irá la lógica para bloquear/desbloquear usuario.");
-                    break;
                 case 0:
-                    System.out.println("↩ Volviendo al menú principal...");
+                    System.out.println("Volviendo al menú principal...");
                     break;
                 default:
-                    System.out.println("✗ Opción no válida.");
+                    System.out.println("Opción no válida.");
                     break;
             }
         } while (opcion != 0);