+++ Exp2_S5/Changelog.md
@@ -1,6 +1,12 @@
 # Historial de cambios
 
-## 22/06/2025 - Creación sistema para administrador
+## 22/06/2025 - Creación nuevos asistentes
+- Se habilitaron las opciones para que el admin pueda gestionar asistentes en el sistema, actualmente las opciones son:
+    - 'AgregarAsistenteMenu': permite la creación de nuevos usuarios con el rol asistente, el sistema verifica que el usuario no este registrado previamente en base al rut.
+    - 'EliminarAsistenteMenu': permite eliminar usuarios con el rol asistente mediante una lista con todos los asistentes que estén registrados.
+- Se modifico la clase `PersistenciaInfo` para soportar la nueva escritura y lectura sobre el archivo **admin.csv**
+
+### 22/06/2025 - Creación sistema para administrador
 - Se implemento el sub paquete `libro` perteneciente al paquete `ui`, en el se gestionan los libros de acuerdo al rol del funcionario:
     - **AgregarLibroMenu.java**: accesible para admin y asistente, permite agregar nuevos libros a la biblioteca.
     - **EliminarLibroMenu.java**: accesible solo para admin, permite eliminar libros de la base de datos siempre y cuando no hayan copias en préstamo activo.